Add C++ implentation of histogramdd
This creates a `histogramdd` operator with overloads matching the `Union`
behaviour used in the functional variant. Moving into C++ is preferred because
it can handle torch function automatically instead of needing to differentiate
between the overloads manually.

This also adds a new return type: `std::tuple<Tensor, std::vector<Tensor>>`. For
which I've updated `wrap` to be completely generic for tuples and removed the
old manual definitions.

ghstack-source-id: adbe5df77fc6d60e5a25eb5dd1a603241f4debdb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74200